+++ crates/cli/src/commands/migrate/mod.rs
@@ -1,8 +1,8 @@
 mod v0_20;
 
+use crate::error::ProtoCliError;
 use clap::Args;
 use starbase::system;
-use starbase_styles::color;
 
 #[derive(Args, Clone, Debug)]
 pub struct MigrateArgs {
@@ -17,10 +17,10 @@ pub async fn migrate(args: ArgsRef<MigrateArgs>) {
             v0_20::migrate().await?;
         }
         unknown => {
-            return Err(miette::miette!(
-                "Unknown migration operation {}.",
-                color::symbol(unknown)
-            ));
+            return Err(ProtoCliError::UnknownMigration {
+                op: unknown.to_owned(),
+            }
+            .into());
         }
     }
 }